Add CDK publishing setup. (#3156)
Allow us to publish CDK versions on PyPi via Github workflows.

This lets us:
- Version CDK moving forward
- Publish via Github actions to both PyPi and test PyPi for test releases.

Note, this will not work on this branch as Github only detects new workflows after they are checked into master, so I was forced to develop and test this from a fork. Browse this PR or this PR to see this command in action.

Added publishing instructions to the Release Document.