Release CDK 0.1.1. (#3360)
Release CDK 0.1.1. Unfortunately I can't release 0.1.0 because we previously released it while testing and I didn't realise once a package is released on PyPi, a package with the same name/version cannot be released regardless of whether it's deleted or not.

Updated:
- source-exchange-rates
- source-github-singer
- source-slack
- source-smartsheets (this is at 0.1.4 since someone published 0.1.3 but the change was not checked in)
- source-stripe

Only source-github-singer and source-smartsheets need to be updated. The rest are either not already published or have singer equivalents present.